Ajouter les modifications pour les catégories

+++ client/src/views/admin/users/AdminUsers.jsx
@@ -6,14 +6,20 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPenToSquare } from '@fortawesome/free-regular-svg-icons';
 import { faTrash } from "@fortawesome/free-solid-svg-icons";
 
+import ValidateModal from "../../../components/modal/ValidateModal";
 import "./adminUsers.scss";
 
 function AdminUsers() {
   document.title = "Back office || Gestion des utilisateurs";
-  const navigate = useNavigate();
   const { session, isLoading } = useSession();
   const [users, setUsers] = useState([]);
   const [shouldRefreshUsers, setShouldRefreshUsers] = useState(false);
+  const [error, setError] = useState(null);
+  const [success, setSuccess] = useState(false);
+  const [showModal, setShowModal] = useState(false);
+  const [userToDelete, setUserToDelete] = useState(null);
+
+  const navigate = useNavigate();
 
   useEffect(() => {
     if (!session) return;
@@ -31,54 +37,72 @@ function AdminUsers() {
           }
         );
 
-        console.log(response);
+        if (!response) {
+          setError("Aucun produits trouvés");
+          setSuccess(false);
+          return;
+        }
 
-        if (!response.ok) return;
-
-        const data = await response.json();
-        console.log("Donnée reçu de l'API:", data);
-        setUsers(data.response);
-        console.log(data.response);
+        if (response.ok) {
+          const data = await response.json();
+          setUsers(data.response);
+          setError(null);
+          setSuccess(false);
+        }
 
       } catch (error) {
-        console.log("Erreur:", error);
+        setError("Erreur de réseau");
+        setSuccess(null);
       }
     }
     fetchUsers();
   }, [session, shouldRefreshUsers]);
 
-
-
-  // Suppression d'un utilisateur
-  async function handleDelete(e, id) {
+  // Affichage de la modal de validation
+  function handleDeleteClick(e, user) {
     e.preventDefault();
     e.stopPropagation();
+    setUserToDelete(user);
+    setShowModal(true);
+  }
+
+  // Suppression d'un utilisateur
+  async function handleValidateDelete() {
+    if (!userToDelete) return;
+
     try {
       const response = await fetch(
-        `http://localhost:9000/api/v1/auth/${id}`,
+        `http://localhost:9000/api/v1/auth/${userToDelete.id}`,
         {
           method: "DELETE",
           credentials: "include",
         }
       );
 
-      console.log(response);
-
       if (!response) {
-        console.log("Erreur lors de la suppression");
-        return;
+        setError("Erreur lors de la suppression");
+        setSuccess(false);
       }
 
       if (response.ok) {
         setShouldRefreshUsers(prev => !prev);
-        console.log("Supprimer avec succès")
-        return;
+        setSuccess(true);
+        setError(null);
       }
-
     } catch (error) {
-      console.log("Erreur:", error);
+      setError(`Erreur système: ${error.message}`);
+      setSuccess(false);
+    } finally {
+      setShowModal(false);
+      setUserToDelete(null);
     }
   }
+
+    // Annuler la suppression
+    function handleCancelDelete() {
+      setShowModal(false);
+      setUserToDelete(null);
+    }
 
   if (isLoading) {
     return (
@@ -97,12 +121,20 @@ function AdminUsers() {
   }
 
   return (
-    <main>
-      <header>
+    <main id="gestion-users">
         <h1>Gestion des utilisateurs</h1>
-        {/* Ajout d'une barre de recherche */}
-      </header>
-      <section>
+
+        {/* Affichage de l'erreur */}
+        {error && <div className="error-message">{error}</div>}
+
+        {/* Affichage du message de succès */}
+        { success && (
+          <div className="success-message">
+            Récupération de tout les utilisateurs avec succès !
+          </div>
+        )}
+
+      <section className="list-users">
         <table>
           <thead>
             <tr>
@@ -122,10 +154,10 @@ function AdminUsers() {
                   <td>{user.email}</td>
                   <td>{user.roles_label}</td>
                   <td>
-                    <button onClick={(e) => { e.stopPropagation(); navigate(`/admin/edituser/${user.id}`); }}>
+                    <button className="btn-edit" onClick={(e) => { e.stopPropagation(); navigate(`/admin/edituser/${user.id}`); }}>
                       <FontAwesomeIcon icon={faPenToSquare} />
                     </button>
-                    <button onClick={(e) => handleDelete(e, user.id)}>
+                    <button className="btn-delete" onClick={(e) => handleDeleteClick(e, user)}>
                       <FontAwesomeIcon icon={faTrash} />
                     </button>
                   </td>
@@ -134,10 +166,19 @@ function AdminUsers() {
             })}
           </tbody>
         </table>
+
         <Link to={"/admin/dashboard"}>Retour au tableau de bord</Link>
       </section>
+
+      {/* Modal de validation */}
+      <ValidateModal
+        show={showModal}
+        message="Êtes-vous sûr de vouloir supprimer cet utilisateur ?"
+        onConfirm={handleValidateDelete}
+        onCancel={handleCancelDelete}
+      />
     </main>
-  )
+  );
 }
 
 export default AdminUsers;
